Added check for difference between decimal and thousands separator settings
I had the problem that I only changed the thousand separator to `.` (it's the German style) but that leads to the trimming of all zeros of the axis labels. For example numbers like `4000` where trimmed to `4`.

This is why I added the check to `Graph::SetNumStringOptions()`, so that this cannot happen anymore.