AxisDateTime: use local timezone instead of UTC
DateTime::__construct() takes two parameters: $time and $timezone. I
will skip the second one as it was not used in AxisDateTime::dateText().
As said in the PHP Manual[1], if $time is a string containing an
‘@’‐prefixed UNIX timestamp, then the current (server‐local) timezone is
ignored and the new DateTime object has UTC set as its timezone. That is
the way dateText() used to work.

For example, if 2020-05-01T12:00+02:00 was passed to SVGGraph (on a
server in the +02:00 timezone), it would be rendered as 2020-05-01
10:00, because that would be the time in UTC. I think it should be
rendered as 2020-05-01 12:00 instead.

This commit replaces the described use of DateTime constructor with a
simple date() call. date() uses the current timezone, not UTC. Users can
override the timezone using date_default_timezone_set().

Links:
[1]: https://www.php.net/manual/en/datetime.construct.php#refsect1-datetime.construct-parameters